types: Improve type support when using None()
Previously invoking None() had a return type of Option<never>. This is
not useful when modeling values that may eventually become a Some() as
the generic for that Option would not be constrained when the value is
updated causing a loss of type support.

This change carries through the generic in a None and better preserves
the type information throughout a system.